backend/requirements.txt
# ============================================================================
# Dependencias Python — Backend Gestión Flujo de Caja
# ============================================================================

# --- Framework web y servidor ASGI ---
fastapi==0.110.3
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# --- Base de datos ---
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# --- Autenticación y seguridad ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
qrcode[pil]==7.4.2
slowapi==0.1.9

# --- Cifrado ---
cryptography==42.0.8

# --- Generación de PDF ---
weasyprint==62.3

# --- Excel (lectura y escritura) ---
openpyxl==3.1.5

# --- Procesamiento de imágenes (firmas Wacom) ---
Pillow==10.3.0

# --- Validación y utilidades ---
pydantic==2.7.4
pydantic-settings==2.3.4
python-dateutil==2.9.0
email-validator==2.1.1

# --- HTTP (para sincronización nodo Bata) ---
httpx==0.27.0

# --- Logging y monitorización ---
structlog==24.2.0
slowapi==0.1.9
bcrypt==4.0.1
